requirements.txt
torch==2.1.2 --index-url https://download.pytorch.org/whl/cpu
torchvision==0.16.2 --index-url https://download.pytorch.org/whl/cpu
Pillow
flask
pandas
gunicorn==20.1.0
pytest==6.2.4
numpy==1.26.2